+++ package.json
@@ -52,7 +52,7 @@
     "dependencies": {
         "@hookform/resolvers": "^3.9.0",
         "@icons-pack/react-simple-icons": "^10.0.0",
-        "@next/third-parties": "^14.2.14",
+        "@next/third-parties": "^14.2.15",
         "@opentelemetry/api": "^1.9.0",
         "@radix-ui/react-accordion": "^1.2.1",
         "@radix-ui/react-alert-dialog": "^1.1.2",
@@ -83,8 +83,8 @@
         "@radix-ui/react-tooltip": "^1.1.3",
         "@supabase/ssr": "^0.5.1",
         "@supabase/supabase-js": "^2.45.4",
-        "@tanstack/react-query": "^5.59.0",
-        "@tanstack/react-query-devtools": "^5.59.0",
+        "@tanstack/react-query": "^5.59.6",
+        "@tanstack/react-query-devtools": "^5.59.6",
         "@tanstack/react-table": "^8.20.5",
         "@vercel/analytics": "^1.3.1",
         "@vercel/otel": "^1.10.0",
@@ -96,8 +96,8 @@
         "date-fns": "^4.1.0",
         "embla-carousel-react": "^8.3.0",
         "input-otp": "^1.2.4",
-        "lucide-react": "^0.447.0",
-        "next": "14.2.14",
+        "lucide-react": "^0.451.0",
+        "next": "14.2.15",
         "next-themes": "^0.3.0",
         "react": "^18",
         "react-day-picker": "^9.1.3",
@@ -118,27 +118,27 @@
         "@commitlint/config-conventional": "^19.5.0",
         "@commitlint/types": "^19.5.0",
         "@eslint/js": "^9.12.0",
-        "@tanstack/eslint-plugin-query": "^5.59.1",
+        "@tanstack/eslint-plugin-query": "^5.59.7",
         "@types/eslint__js": "^8.42.3",
         "@types/node": "^22",
         "@types/react": "^18",
         "@types/react-dom": "^18",
         "@types/validator": "^13.12.2",
         "eslint": "^8.57.0",
-        "eslint-config-next": "14.2.14",
+        "eslint-config-next": "14.2.15",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-perfectionist": "^3.8.0",
         "eslint-plugin-prettier": "^5.2.1",
         "eslint-plugin-react": "^7.37.1",
-        "globals": "^15.10.0",
+        "globals": "^15.11.0",
         "husky": "^9.1.6",
         "lint-staged": "^15.2.10",
         "postcss": "^8",
         "prettier": "^3.3.3",
         "prettier-plugin-tailwindcss": "^0.6.8",
-        "supabase": "^1.200.3",
+        "supabase": "^1.204.3",
         "tailwindcss": "^3.4.13",
-        "typescript": "^5.6.2",
+        "typescript": "^5.6.3",
         "typescript-eslint": "^8.8.1"
     }
 }
